// src/app/types/types.ts
export interface IRouteItem {
  id: number;
  value: string;
  link: string;
}
export interface IAccordionItem {
  id: number;
  title: string;
  content: string;
}
export interface ISliderRoute {
  id: number;
  value: string;
  price: number;
  link: string;
  src: string;
}